agent/pyproject.toml
[project]
name = "meeting-assistant-agent"
version = "0.1.0"
description = "AI-агент для работы с корпоративными созвонами через Follow-Up"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "a2a-sdk[http-server]>=0.3.4",
    "starlette>=0.38.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]
telemetry = [
    "openinference-instrumentation-langchain>=0.1.0",
    "arize-phoenix-otel>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W"]







